+++ apps/ledger-live-desktop/src/renderer/components/TrackAppStart.tsx
@@ -4,6 +4,6 @@ import { hasCompletedOnboardingSelector } from "~/renderer/reducers/settings";
 import Track from "~/renderer/analytics/Track";
 const TrackAppStart = () => {
   const hasCompletedOnboarding = useSelector(hasCompletedOnboardingSelector);
-  return hasCompletedOnboarding ? <Track mandatory onMount event="App Starts" /> : null;
+  return hasCompletedOnboarding ? <Track onMount event="App Starts" /> : null;
 };
 export default TrackAppStart;