+++ libs/ledger-live-common/src/families/tron/serialization.ts
@@ -1,6 +1,15 @@
 import { BigNumber } from "bignumber.js";
-import type { TronAccount, TronAccountRaw, TronResources, TronResourcesRaw } from "./types";
-import { Account, AccountRaw } from "@ledgerhq/types-live";
+import {
+  isTrongridExtraTxInfo,
+  isTrongridExtraTxInfoRaw,
+  type TronAccount,
+  type TronAccountRaw,
+  type TronResources,
+  type TronResourcesRaw,
+  type TrongridExtraTxInfo,
+  type TrongridExtraTxInfoRaw,
+} from "./types";
+import { Account, AccountRaw, OperationExtra, OperationExtraRaw } from "@ledgerhq/types-live";
 
 export const toTronResourcesRaw = ({
   frozen,
@@ -21,8 +30,10 @@ export const toTronResourcesRaw = ({
   const delegatedFrozenBandwidth = delegatedFrozen.bandwidth;
   const delegatedFrozenEnergy = delegatedFrozen.energy;
   const cacheTransactionInfoById = {};
-  const legacyFrozenBandwidth = legacyFrozen.bandwidth;
-  const legacyFrozenEnergy = legacyFrozen.energy;
+  const unFrozenBandwidth = unFrozen?.bandwidth;
+  const unFrozenEnergy = unFrozen?.energy;
+  const legacyFrozenBandwidth = legacyFrozen?.bandwidth;
+  const legacyFrozenEnergy = legacyFrozen?.energy;
 
   for (const k in cacheTx) {
     const { fee, blockNumber, withdraw_amount, unfreeze_amount } = cacheTx[k];
@@ -55,13 +66,13 @@ export const toTronResourcesRaw = ({
         : undefined,
     },
     unFrozen: {
-      bandwidth: unFrozen.bandwidth
-        ? unFrozen.bandwidth.map(entry => {
+      bandwidth: unFrozenBandwidth
+        ? unFrozenBandwidth.map(entry => {
             return { amount: entry.amount.toString(), expireTime: entry.expireTime.toISOString() };
           })
         : undefined,
-      energy: unFrozen.energy
-        ? unFrozen.energy.map(entry => {
+      energy: unFrozenEnergy
+        ? unFrozenEnergy.map(entry => {
             return { amount: entry.amount.toString(), expireTime: entry.expireTime.toISOString() };
           })
         : undefined,
@@ -115,8 +126,10 @@ export const fromTronResourcesRaw = ({
   const frozenEnergy = frozen.energy;
   const delegatedFrozenBandwidth = delegatedFrozen.bandwidth;
   const delegatedFrozenEnergy = delegatedFrozen.energy;
-  const legacyFrozenBandwidth = legacyFrozen.bandwidth;
-  const legacyFrozenEnergy = legacyFrozen.energy;
+  const unFrozenBandwidth = unFrozen?.bandwidth;
+  const unFrozenEnergy = unFrozen?.energy;
+  const legacyFrozenBandwidth = legacyFrozen?.bandwidth;
+  const legacyFrozenEnergy = legacyFrozen?.energy;
 
   const cacheTransactionInfoById = {};
 
@@ -158,13 +171,13 @@ export const fromTronResourcesRaw = ({
         : undefined,
     },
     unFrozen: {
-      bandwidth: unFrozen.bandwidth
-        ? unFrozen.bandwidth.map(entry => {
+      bandwidth: unFrozenBandwidth
+        ? unFrozenBandwidth.map(entry => {
             return { amount: new BigNumber(entry.amount), expireTime: new Date(entry.expireTime) };
           })
         : undefined,
-      energy: unFrozen.energy
-        ? unFrozen.energy.map(entry => {
+      energy: unFrozenEnergy
+        ? unFrozenEnergy.map(entry => {
             return { amount: new BigNumber(entry.amount), expireTime: new Date(entry.expireTime) };
           })
         : undefined,
@@ -212,4 +225,62 @@ export function assignFromAccountRaw(accountRaw: AccountRaw, account: Account) {
   const tronResourcesRaw = (accountRaw as TronAccountRaw).tronResources;
   if (tronResourcesRaw)
     (account as TronAccount).tronResources = fromTronResourcesRaw(tronResourcesRaw);
+}
+
+export function fromOperationExtraRaw(extraRaw: OperationExtraRaw): TrongridExtraTxInfo {
+  const extra: TrongridExtraTxInfo = {};
+  if (!isTrongridExtraTxInfoRaw(extraRaw)) {
+    return extra;
+  }
+
+  if (extraRaw.frozenAmount) {
+    extra.frozenAmount = new BigNumber(extraRaw.frozenAmount);
+  }
+
+  if (extraRaw.unfreezeAmount) {
+    extra.unfreezeAmount = new BigNumber(extraRaw.unfreezeAmount);
+  }
+
+  if (extraRaw.votes) {
+    extra.votes = extraRaw.votes;
+  }
+
+  if (extraRaw.unDelegatedAmount) {
+    extra.unDelegatedAmount = new BigNumber(extraRaw.unDelegatedAmount);
+  }
+
+  if (extraRaw.receiverAddress) {
+    extra.receiverAddress = extraRaw.receiverAddress;
+  }
+
+  return extra;
+}
+
+export function toOperationExtraRaw(extra: OperationExtra): TrongridExtraTxInfoRaw {
+  const extraRaw: TrongridExtraTxInfoRaw = {};
+  if (!isTrongridExtraTxInfo(extra)) {
+    return extraRaw;
+  }
+
+  if (extra.frozenAmount) {
+    extraRaw.frozenAmount = extra.frozenAmount.toString();
+  }
+
+  if (extra.unfreezeAmount) {
+    extraRaw.unfreezeAmount = extra.unfreezeAmount.toString();
+  }
+
+  if (extra.votes) {
+    extraRaw.votes = extra.votes;
+  }
+
+  if (extra.unDelegatedAmount) {
+    extraRaw.unDelegatedAmount = extra.unDelegatedAmount.toString();
+  }
+
+  if (extra.receiverAddress) {
+    extraRaw.receiverAddress = extra.receiverAddress;
+  }
+
+  return extraRaw;
 }